Regression suite builds using make
reworked and cleaned up so that a change to any of the files in the
dependencies list will cause dependencies to regenerate.

Still a slight clitch in that the make file is regenerated more often than it
should need to, but in practice that is not causing a significant problem.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>