HPCC-13173 Constant folding should not call plugin functions
Only specifically whitelisted functions should be constant-folded.

Note that this does not prevent people from declaring their own services and
marking the functions in them as ,fold, so it does not close the security
loophole just yet - that will require some work on the code signing issue
HPCC-14185 - but this is a necessary step towards it (and a useful feature
in its own right).

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>